Updated the pFL-SimCLR based on the ssl platform
